Refactor endernodetracker to be in combat.endernodetracker
