+++ src/main/java/at/hannibal2/skyhanni/utils/CombatUtils.kt
@@ -1,7 +1,7 @@
 package at.hannibal2.skyhanni.utils
 
-import at.hannibal2.skyhanni.features.misc.ghostcounter.GhostCounter
-import at.hannibal2.skyhanni.features.misc.ghostcounter.GhostData
+import at.hannibal2.skyhanni.features.combat.ghostcounter.GhostCounter
+import at.hannibal2.skyhanni.features.combat.ghostcounter.GhostData
 import io.github.moulberry.notenoughupdates.core.util.lerp.LerpUtils
 import io.github.moulberry.notenoughupdates.util.XPInformation
